mouse lock with F1, mouse ocords float.

+++ src/common/arcball_camera.h
@@ -19,15 +19,15 @@ struct Camera
     float near_plane = 1.0f;
     float far_plane = 5.0f;
     float fovy_degrees = 45.0f;
-    int _mx = 320;
-    int _my = 240;
+    float _mx = 320.0f;
+    float _my = 240.0f;
 
     mat4 proj;
     mat4 view;
 
     virtual void translate(const glm::vec3 &);
     virtual void mouse_click(int mx, int my);
-    virtual void mouse_move(int mx, int my);
+    virtual void mouse_move(float mx, float my);
     virtual void update();
 };
 
@@ -40,7 +40,7 @@ struct FpsCamera : Camera
     vec3 dir = vec3(0,0,-1);
 
     void mouse_click(int mx, int my) override;
-    void mouse_move(int mx, int my) override;
+    void mouse_move(float mx, float my) override;
     void update() override;
 };
 
@@ -81,7 +81,7 @@ struct ArcballCamera : Camera
     void mouse_click(int mx, int my) override;
 
     // on move the grabbed point of the sphere.
-    void mouse_move(int mx, int my) override;
+    void mouse_move(float mx, float my) override;
 
 
     void update() override;
